+++ test/mocha.opts
@@ -1,0 +1,3 @@
+--ui bdd
+--compilers js:babel-register
+--recursive
