+++ test/specs/init_cmd.spec.js
@@ -1,5 +1,5 @@
 import assert from 'assert'
-import initCmd from '../src/init_cmd'
+import initCmd from '../../src/init_cmd'
 
 describe('initCmd', () => {
 	it('should be a function', () => {
