+++ test/specs/shell_cmd.spec.js
@@ -1,8 +1,8 @@
 import assert from 'assert'
 import sinon from 'sinon'
-import shellCmd from '../src/shell_cmd'
+import shellCmd from '../../src/shell_cmd'
 import child_process from 'child_process'
-import * as config from '../src/config'
+import * as config from '../../src/config'
 
 describe('shellCmd', () => {
 	const sandbox = sinon.sandbox.create()
